requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]==3.6
scikit-learn 
pyaml
dill 
dagshub
mlflow==2.7.1
fastapi
python-multipart

#-e .